+++ kibot/globals.py
@@ -238,7 +238,8 @@ class Globals(FiltersOptions):
             """ When applying filters and variants remove the solder mask apertures for components that won't be included """
             self.restore_project = False
             """ Restore the KiCad project after execution.
-                Note that this option will undo operations like `set_text_variables` """
+                Note that this option will undo operations like `set_text_variables`.
+                Starting with 1.6.4 it also restores the PRL (Project Local Settings) file """
             self.set_text_variables_before_output = False
             """ Run the `set_text_variables` preflight before running each output that involves variants.
                 This can be used when a text variable uses the variant and you want to create more than